imageAnimation/requirements.txt
google-resumable-media==0.6.0
google-cloud-storage==1.30.0
google-cloud-bigquery==1.26.1
pandas
numpy==1.21.6
scikit-learn==0.22